Refactor error handling and constants for optimization
