+++ src/training/steps/pre_training/feature_lookback_optimization/core/vectorbt_scoring.py
@@ -56,7 +56,7 @@ except ImportError:
 
 from src.utils.tprint import tprint, tprint_error, tprint_success, tprint_warning, tprint_debug, tprint_info
 from src.utils.logger import get_logger
-from .utils.error_handling import safe_operation, get_error_handler
+from ..error_handling.error_handler import safe_operation, get_error_handler
 
 logger = get_logger('VectorBTScoring')
