+++ src/training/steps/pre_training/feature_lookback_optimization/error_handling/__init__.py
@@ -7,8 +7,23 @@ This module provides standardized error handling for feature lookback optimizati
 # Import centralized tprint utilities
 from ..utils.tprint_utils import tprint_debug, TPRINT_AVAILABLE
 
-from .error_handler import StandardizedErrorHandler, ErrorSeverity, ErrorCategory, ErrorDetails
-# ErrorRecoveryResult doesn't exist - removed from import
+from .error_handler import (
+    StandardizedErrorHandler, 
+    ErrorSeverity, 
+    ErrorCategory,
+    ErrorDetails,
+    OptimizationError,
+    DataValidationError,
+    ScoringError,
+    CacheError,
+    MemoryError,
+    safe_operation,
+    safe_mi_calculation,
+    safe_correlation_calculation,
+    safe_dataframe_operation,
+    safe_numpy_operation,
+    get_error_handler
+)
 
 # Log module import
 if TPRINT_AVAILABLE:
@@ -18,5 +33,16 @@ __all__ = [
     'StandardizedErrorHandler',
     'ErrorSeverity',
     'ErrorCategory',
-    'ErrorDetails'
+    'ErrorDetails',
+    'OptimizationError',
+    'DataValidationError',
+    'ScoringError',
+    'CacheError',
+    'MemoryError',
+    'safe_operation',
+    'safe_mi_calculation',
+    'safe_correlation_calculation',
+    'safe_dataframe_operation',
+    'safe_numpy_operation',
+    'get_error_handler'
 ]
