+++ src/training/steps/pre_training/feature_lookback_optimization/utils/fast_failing_validation.py
@@ -23,7 +23,7 @@ from .tprint_utils import (
     TPRINT_AVAILABLE
 )
 
-from .error_handling import DataValidationError, OptimizationError
+from ..error_handling.error_handler import DataValidationError, OptimizationError
 from .nan_handling import SafeNaNHandler
